feat(frida): separate Frida Setup and Scripts into distinct screens
Split the Frida script editor out of the Frida server setup screen into a
dedicated "Frida Scripts" screen and menu item. Redesign the scripts list
(header with white "Add Script" button + dropdown for new/upload, search bar
with icon, script cards showing icon, name, timestamp and content preview) and
add a confirmation dialog before deleting a script.

- Add FridaScriptsComposable with list, search, create/upload and editor flows
- Trim FridaComposable down to the Frida server setup only
- Register the new screen and menu entry in Navigation
- Add ic_code and search_icon drawables
